+++ examples/complete/provider.tf
@@ -1,4 +1,11 @@
 provider "aws" {
   profile = "default"
   region  = "us-east-1"
+
+  default_tags {
+    tags = {
+      ManagedBy = "Terraform"
+      CreatedAt = timestamp()
+    }
+  }
 }